+++ apps/web/src/app/vault/individual-vault.component.ts
@@ -25,8 +25,6 @@ import { SyncService } from "@bitwarden/common/abstractions/sync.service";
 import { TokenService } from "@bitwarden/common/abstractions/token.service";
 import { CipherView } from "@bitwarden/common/models/view/cipherView";
 
-import { VaultFilterComponent } from "../modules/vault-filter/vault-filter.component";
-import { VaultFilterService } from "../modules/vault-filter/vault-filter.service";
 import { UpdateKeyComponent } from "../settings/update-key.component";
 
 import { AddEditComponent } from "./add-edit.component";
@@ -36,6 +34,8 @@ import { CollectionsComponent } from "./collections.component";
 import { FolderAddEditComponent } from "./folder-add-edit.component";
 import { ShareComponent } from "./share.component";
 import { VaultService } from "./shared/vault.service";
+import { VaultFilterService } from "./vault-filter/shared/vault-filter.service";
+import { VaultFilterComponent } from "./vault-filter/vault-filter.component";
 
 const BroadcasterSubscriptionId = "VaultComponent";
