+++ apps/web/src/app/vault/vault-filter/shared/folder-filter/folder-filter.component.ts
